requirements.txt
selenium >= 4.3.0
robotframework >= 4.1.3
robotframework-pythonlibcore >= 4.4.1
robotframework-seleniumlibrary >= 5.1.1
robotframework-browser
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-api-keys
Faker
beautifulsoup4
yagmail
dominate
